Added Program and program_course entity with provider association and Liquibase migration

+++ src/main/java/com/dmantz/lms_b/service/impl/CourseManagementServiceImpl.java
@@ -274,8 +274,8 @@ public class CourseManagementServiceImpl implements CourseManagementService {
 					throw new DuplicateValuesException("Chapter name already exists in this course");
 				});
 
-		int nextChapterNum = chapterRepository.findTopByCourse_CourseIdOrderByChapterNumDesc(request.getCourseId())
-				.map(ch -> ch.getChapterNum() + 1).orElse(1);
+		Long nextChapterNum = chapterRepository.findTopByCourse_CourseIdOrderByChapterNumDesc(request.getCourseId())
+				.map(ch -> ch.getChapterNum() + 1).orElse(1L);
 		Chapter chapter = chapterMapper.toEntity(request);
 		chapter.setCourse(course); // ✅ managed entity
 		chapter.setChapterNum(nextChapterNum);
